refactor: split large builder files into focused submodules
Convert four flat files >500 lines into directory modules for better
maintainability:

- src/aggregate/builder.rs → builder/{mod,single,set,tests}.rs
  AggregateFunctionBuilder in single.rs, AggregateFunctionSetBuilder +
  OverloadBuilder in set.rs, 14 unit tests in tests.rs

- src/scalar/builder.rs → builder/{mod,single,set,tests}.rs
  ScalarFn + ScalarFunctionBuilder in single.rs, ScalarFunctionSetBuilder +
  ScalarOverloadBuilder in set.rs, 13 unit tests in tests.rs

- src/scaffold/mod.rs: extract templates.rs (11 pub(super) generate_* fns)
  and tests.rs (29 unit tests); mod.rs retains public API only

- src/table/builder.rs: surgically extract BindInfo, InitInfo, FunctionInfo
  into new src/table/info.rs; update table/mod.rs re-exports accordingly

Update CONTRIBUTING.md, docs/architecture.md, and book/src/contributing.md
with accurate directory trees reflecting all new submodule structures.

All 303 unit + 40 integration + 90 doc tests pass.
